tests/wpt/web-platform-tests/cookies/path/no-match-no-cookie-trailing-slash.sub.html

<!doctype html>
<html>
<head>
  <meta charset=utf-8>
  <title>Does not set or send cookie from `Set-Cookie` HTTP header for path=/w</title>
  <meta name=help href="http://tools.ietf.org/html/rfc6265#section-5.1.4">
  <script src="/cookies/resources/es6-promise.js"></script>
  <script src="/cookies/resources/fetch.js"></script>

  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/cookies/resources/testharness-helpers.js"></script>
</head>
<body>
<div id=log></div>
<script>
  function clearKnownCookie() {
    document.cookie = "path_matches=0; expires=Thu, 01 Jan 1970 00:00:01 GMT; path=/w";
  }

  test(function () {
      assert_equals(document.cookie.match(/path_matches=1/), null);
  }, "'path_matches=1' not present in `document.cookie`");

  promise_test(function (t) {
      t.add_cleanup(clearKnownCookie);

      document.cookie = "path_matches=1; expires=Thu, 01 Jan 1970 00:00:01 GMT; path=/w";

      return fetch("/cookies/resources/echo-json.py", { "credentials": "include" })
      .then(function (r) {
        return r.json();
      })
      .then(function (j) {
        assert_equals(j["path_matches"], undefined);
      });
  }, "cookie with path=/w not sent in HTTP request");
</script>
</body>
</html>
